# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for bmstu dormitory app.",
        "title": "BMSTU-WEB API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
...
        "/api/v1/students/{stud-number}/things/": {
            "get": {
                "description": "View full information about all current things of student.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get student things",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student Number",
                        "name": "stud-number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ThingFullInfo"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустым.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Студент не найден!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/things/": {
            "get": {
                "description": "View full information about all things dormitory.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all things in dormitory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ThingFullInfo"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new thing with params in base.",
                "produces": [
                    "application/json"
                ],
                "summary": "Add new thing",
                "parameters": [
                    {
                        "description": "body for buy service",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddNewThingRequestMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция успешно проведена!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Вещь с таким же уникальным номером уже есть в базе",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/things/free/": {
            "get": {
                "description": "View full information about free things (things without owner) .",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all free things in dormitory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ThingFullInfo"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/things/{mark-number}/": {
            "patch": {
                "description": "Transfer thing to another room.",
                "produces": [
                    "application/json"
                ],
                "summary": "Transfer thing to another room.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Thing mark number",
                        "name": "mark-number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dst room in which thing will be transferred.",
                        "name": "room-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransferThingRequestMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вещь успешно перемещена!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Параметр не должен быть пустой\" | \"Параметр должен быть числом!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Вещь не найдена\" | \"Комната не найдена",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "422": {
                        "description": "Вещь уже находится в этой комнате!",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Проблемы на стороне сервера.",
                        "schema": {
                            "$ref": "#/definitions/models.ShortResponseMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.AddNewThingRequestMessage": {
            "type": "object",
            "properties": {
                "markNumber": {
                    "type": "integer"
                },
                "thingType": {
                    "type": "string"
                }
            }
        },
...
        "models.ShortResponseMessage": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "models.ThingFullInfo": {
            "type": "object",
            "properties": {
                "room": {
                    "type": "object",
                    "$ref": "#/definitions/objects.Room"
                },
                "thing": {
                    "type": "object",
                    "$ref": "#/definitions/objects.Thing"
                }
            }
        },
        "models.TransferThingRequestMessage": {
            "type": "object",
            "properties": {
                "room-id": {
                    "type": "integer"
                }
            }
        },
        "objects.Room": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "roomNumber": {
                    "type": "integer"
                },
                "roomType": {
                    "type": "string"
                }
            }
        },
...
        "objects.Thing": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "markNumber": {
                    "type": "integer"
                },
                "ownerID": {
                    "type": "integer"
                },
                "roomID": {
                    "type": "integer"
                },
                "thingType": {
                    "type": "string"
                }
            }
        },
...
    }
}